Removed username from Utente.
Removed username from utente. Now the login is performed by email and
password, rather than by username and password.
The test highlighted a bug due to the char replacement performed by PHP
to the $_GET and $_POST variable to keep compatibility with variable
names. To fix the issue, the query string is parsed manually starting from
$_SERVER['QUERY_STRING'].