Abilitazione el.s of Gara ordered by id and finoAl
In getting the Gara elements, the Abilitazione elements are ordered by id
(of the inner element) ascendent and by FinoAl ascendent. In this way it
is possible to get the current elements by iterating them ant taking the
first elements that has FinoAl greater than now, for each different id.